added the steering vector calcuation function
